==> notebooks/05_model_deploy_sagemaker.ipynb <==
**** Purpose: Deployment to endpoint

#Step0 Directory setups:
#!mkdir -p code
#!mv inference.py code/
#!rm -r bert
!rm -rf code/.ipynb_checkpoints


from transformers import DistilBertTokenizer, DistilBertModel

DistilBertTokenizer.from_pretrained("distilbert-base-uncased").save_pretrained("./bert")
DistilBertModel.from_pretrained("distilbert-base-uncased").save_pretrained("./bert")


#Step 1: Create new tar.gz with code/ included
!tar -czvf model.tar.gz logreg_model.joblib label_encoder.joblib bert code/


!tar -tzvf model.tar.gz

from sagemaker.sklearn.model import SKLearnModel
from sagemaker import get_execution_role
import sagemaker

role = get_execution_role()
sagemaker_session = sagemaker.Session()

#STEP 2: Upload tar to S3
import boto3

s3 = boto3.client('s3')
s3.upload_file(
    Filename='model.tar.gz',
    Bucket='complaint-classifier-jp2025',
    Key='models/model.tar.gz'
)


#Step 3: Deploy using Model()
from sagemaker.model import Model

image_uri = '763104351884.dkr.ecr.us-east-1.amazonaws.com/huggingface-pytorch-inference:1.13.1-transformers4.26.0-cpu-py39-ubuntu20.04'

model = Model(
    image_uri=image_uri,
    model_data="s3://complaint-classifier-jp2025/models/model.tar.gz",  # must contain joblib, bert/, code/
    role=role,
    entry_point="inference.py",
    source_dir="code",
    sagemaker_session=sagemaker_session,
    env={
        "SAGEMAKER_PROGRAM": "inference.py"
    }
)


predictor = model.deploy(
    instance_type='ml.m5.large',
    initial_instance_count=1,
    endpoint_name='complaint-logreg-bert-endpoint'
)


print(predictor)

from sagemaker.predictor import Predictor
from sagemaker.serializers import JSONSerializer

predictor = Predictor(
    endpoint_name="complaint-logreg-bert-endpoint",
    serializer=JSONSerializer()
)

payload = {"text": "I was charged extra interest after my loan was closed."}
response = predictor.predict(payload)
print(response)


#predictor.delete_endpoint()
